kafka-canary-deployer: Rewrite deployer to work with deploying topic to separate cluster
At NAV, the Topic resource must be deployed to the GCP clusters, even for a canary running in FSS clusters.
This rewrite makes it possible to specify where the topic should be deployed, independent of where the canary is deployed.
In addition, the script now does deploys in parallel, nicer logging, and supports dry-run.